Refactor parser into module
